smdk4412-common: camera: enable ZSL mode on front camera
We are querying FIMC2 for the capture output and hence we
have no reason to restart the preview to have the preview
resolution match the capture resolution, which is not how
Zero Shutter Lag works. We could have different resolutions
for the preview and capture, hence don't stop the preview
for resizing.

Change-Id: Iaf2b4d0c51dbedb9516b7c3cdde09aa18cd95f73

Patchset: 1
http://review.cyanogenmod.org/#/c/54817/